Projet_7/dashboard_docker/requirements.txt
altair
argh
astor
attrs
base58
blinker
boto3
botocore
certifi
chardet
Click
cycler
decorator
docutils
entrypoints
enum-compat
Flask
future
idna
imbalanced-learn
imblearn
importlib-metadata
ipython
itsdangerous
Jinja2
jmespath
joblib
jsonschema
kiwisolver
MarkupSafe
MLflow
matplotlib
more-itertools
numpy
pandas
pathtools
Pillow
protobuf
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
requests
s3transfer
scikit-learn
scipy
seaborn
shap 
six
streamlit
toml
toolz
tornado
tzlocal
urllib3
validators
watchdog
Werkzeug
xgboost
zipp
